+++ src/spindry/__init__.py
@@ -2,10 +2,10 @@
 
 from mchammer import Atom, Bond, Molecule
 
-from spindry.potential import SpdPotential, VaryingEpsilonPotential
-from spindry.spinner import Spinner
-from spindry.supramolecule import SupraMolecule
-from spindry.utilities import (
+from spindry._internal.potential import SpdPotential, VaryingEpsilonPotential
+from spindry._internal.spinner import Spinner
+from spindry._internal.supramolecule import SupraMolecule
+from spindry._internal.utilities import (
     calculate_centroid_distance,
     calculate_min_atom_distance,
     get_atom_distance,
