+++ src/spindry/_internal/utilities.py
